refactor: extract Home Assistant state
