+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/). This project use
 - Split cloud authentication, Pecron MQTT callbacks/recovery, poll-rate validation, and the main monitor lifecycle into `monitor_cloud.py` while retaining one facade-owned runtime state (#66).
 - Split control routing, value probing, convenience commands, and Home Assistant command parsing into `monitor_controls.py`, preserving local/cloud fallback order and readback behavior (#66).
 - Split Home Assistant entity policy, exact MQTT discovery catalogs, stale-topic cleanup, and deferred port discovery into `ha_discovery.py` without changing published topics or payloads (#66).
+- Split Home Assistant telemetry formatting, last-known-state handling, deferred-port updates, and energy-counter publishing into `ha_state.py`, leaving `ha_bridge.py` focused on MQTT connection, retry, and command dispatch (#66).
 
 ### Fixed
 - Continuous local monitoring now retries incomplete E3600/E3800 multi-packet reads within the current poll cycle, targets only devices still missing telemetry, and advances to the next future cycle boundary when retries overrun an interval (#88).
