+++ pyproject.toml
@@ -34,6 +34,7 @@ py-modules = [
     "energy_state",
     "ha_bridge",
     "ha_discovery",
+    "ha_state",
     "helpers",
     "lan_scan",
     "local_transport",
@@ -79,6 +80,7 @@ source = [
     "energy_state",
     "ha_bridge",
     "ha_discovery",
+    "ha_state",
     "helpers",
     "lan_scan",
     "local_transport",
